requirements.txt
tweepy>=4.14.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pillow>=10.0.0
openai>=1.3.0
pyyaml>=6.0.1

